Calculate PI - change return type to BigDecimal, and made a little visual refactoring in runner class.

+++ algorithm/src/main/java/com/example/pi/PiCalculator.java
@@ -1,8 +1,10 @@
 package com.example.pi;
 
+import java.math.BigDecimal;
+
 @FunctionalInterface
 public interface PiCalculator {
 
-    double calculate(int n);
+    BigDecimal calculate(int n);
 
 }
